requirements.txt
line-bot-sdk==3.19.1
fastapi
uvicorn[standard]
google-adk
aiohttp
python-dotenv
pymongo
pydantic-settings